+++ src/general/reset.py
@@ -1,7 +1,7 @@
 """Provides classes for general Reset"""
 
-from src.commands import Commands, ResultAndError
-from src.packet import Packet, PacketAck
+from src.protocol.commands import Commands, ResultAndError
+from src.protocol.packet import Packet, PacketAck
 
 class PacketGeneralReset(Packet):
     """Packet for general Reset"""
