+++ src/protocol/protocol.py
@@ -1,13 +1,13 @@
 """Provides helper class for ScienceMode protocol"""
 
 import asyncio
-from src.commands import Commands
+from src.protocol.commands import Commands
 from src.general.general_error import PacketGeneralError
 from src.general.unknown_command import PacketGeneralUnknownCommand
-from src.packet_factory import PacketFactory
+from src.protocol.packet_factory import PacketFactory
 from src.utils.byte_builder import ByteBuilder
 from src.utils.crc16 import Crc16
-from src.packet import Packet, PacketAck
+from src.protocol.packet import Packet, PacketAck
 from src.utils.connection import Connection
